requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
groq>=0.12.0
requests>=2.31.0
python-dotenv>=1.0.0
psutil>=6.0.0
pygetwindow>=0.0.9
pyperclip>=1.9.0
pyautogui>=0.9.54
playwright>=1.48.0
SpeechRecognition>=3.10.0
websocket-client>=1.6.0
PyAudio>=0.2.13
